refactor: Remove unused CSS styles and update RouterProvider layout
